// apps/public-portal/src/app/core/public-api/dtos/institution-section-payloads.dto.ts
export interface InstitutionHeroSectionPayloadDto {
  imageUrl?: string;
  badge?: string;
  title?: string;
  description?: string;
  applyNowLabel?: string;
  applyNowTrustUrl?: string;
}

export interface InstitutionOverviewSectionPayloadDto {
  sectionLabel?: string;
  title?: string;
  description?: string;
}

export interface InstitutionCoursesIntroSectionPayloadDto {
  sectionLabel?: string;
  title?: string;
  description?: string;
}

export interface InstitutionScholarshipsIntroSectionPayloadDto {
  sectionLabel?: string;
  title?: string;
  description?: string;
}